Expunge all the excel reader code to save memory in prod.
Summary:
khanacademy.org/stats/memory shows that one of the biggest users of
memory is the openpxyl.reader.iter_worksheet module, because it
creates this giant table that maps column-name to column-number.  (And
maybe for other reasons as well, I didn't look too deeply:)

   name                           bytes
   instance_cache                 25,279,200
   gae_bingo.instance_cache        4,598,488
   openpyxl.reader.iter_worksheet  1,140,280
   api.api_request_class           1,100,232
   linecache                       1,078,432

There it is at #3!

We don't even use this class, because we never try to read
spreadsheets, we just write them.  So I'm nixing the whole reader/
subdirectory.  If we ever need that functionality we can undelete it
again.

Test Plan: tools/runtests.py coach_resources

Reviewers: mattfaus, chris

Reviewed By: mattfaus

Differential Revision: http://phabricator.khanacademy.org/D6656